As first approach to speed up the chunk 3 reading, I've added a pxd file so cython can use static types.

Usage: run `cythonize -3 -i -a pyxdf.py` in the pyxdf subfolder, import pyxdf as usual.

So far, it yields a speedup of about 16% (3.2s -> 2.7s), whereas my previous changes combined reduced the time from about 7s to 3.83s, so I'm not sure if it's worth it without replacing some functions with cython code.

**Edit**: It works with the python file classes, so gzip support is still there.