@tstenner 2020-02-24
This has become usable, has the same interface and almost no performance impact without the compiled extension (±.5%), so it could be merged.