@cbrnr 2020-02-24
Very nice, that's a good start! Can you add a CHANGELOG entry?